+++ src/tika_client/data_models.py
@@ -1,11 +1,11 @@
 from datetime import datetime
 from enum import Enum
 from typing import Dict
-from typing import Final
 from typing import List
 from typing import Optional
-from typing import Type
 from typing import Union
+
+# Based on https://cwiki.apache.org/confluence/display/TIKA/Metadata+Overview
 
 
 class TikaKey(str, Enum):
@@ -13,10 +13,36 @@ class TikaKey(str, Enum):
     ContentType = "Content-Type"
     ContentLength = "Content-Length"
     Content = "X-TIKA:content"
+
+
+class DublinCoreKey(str, Enum):
+    Creator = "dc:creator"
     Created = "dcterms:created"
     Modified = "dcterms:modified"
+    Rights = "dc:rights"
+    Contributor = "dc:contributor"
     Title = "dc:title"
+    Relation = "dc:relation"
+    Type = "dc:type"
+    Identifier = "dc:identifier"
+    Publisher = "dc:publisher"
+    Description = "dc:description"
     Subject = "dc:subject"
+    Language = "dc:language"
+    Format = "dc:format"
+
+
+class XmpKey(str, Enum):
+    About = "xmp:About"
+    Created = "xmp:CreateDate"
+    NumPages = "xmpTPg:NPages"
+
+
+class OtherTikaKeys(str, Enum):
+    CharacterCount = "meta:character-count"
+    LastAuthor = "meta:last-author"
+    Revision = "cp:revision"
+    Language = "language"
 
 
 class TikaResponse:
@@ -30,17 +56,38 @@ class TikaResponse:
 
     def __init__(self, data: Dict) -> None:
         self.data = data
+        # Always set keys
         self.type: str = self.data[TikaKey.ContentType]
         self.parsers: List[str] = self.data[TikaKey.Parsers]
 
+        # Tika keys
+        self.content = self.get_optional_string(TikaKey.Content)
+        self.content_length = self.get_optional_int(TikaKey.ContentLength)
+
+        # Dublin Core keys
+        self.created = self.get_optional_datetime(DublinCoreKey.Created)
+        self.modified = self.get_optional_datetime(DublinCoreKey.Modified)
+        self.title = self.get_optional_string(DublinCoreKey.Title)
+
+        # Xmp keys
+        # TODO: Implement more of these
+        self.xmp_created = self.get_optional_datetime(XmpKey.Created)
+        self.page_count = self.get_optional_int(XmpKey.NumPages)
+
+        # Other general keys
+        self.character_count = self.get_optional_int(OtherTikaKeys.CharacterCount)
+        self.revision = self.get_optional_int(OtherTikaKeys.Revision)
+        self.language = self.get_optional_string(OtherTikaKeys.Language)
+        self.last_author = self.get_optional_string(OtherTikaKeys.LastAuthor)
+
     # Helpers
 
-    def get_optional_int(self, key: Union[TikaKey, str]) -> Optional[int]:
+    def get_optional_int(self, key: Union[TikaKey, DublinCoreKey, XmpKey, str]) -> Optional[int]:
         if key not in self.data:  # pragma: no cover
             return None
         return int(self.data[key])
 
-    def get_optional_datetime(self, key: Union[TikaKey, str]) -> Optional[datetime]:
+    def get_optional_datetime(self, key: Union[TikaKey, DublinCoreKey, XmpKey, str]) -> Optional[datetime]:
         """
         If present, attempts to parse the given key as an ISO-8061 format
         datetime, including timezone handling and return if.
@@ -52,93 +99,10 @@ class TikaResponse:
         # Handle Zulu time as UTC
         return datetime.fromisoformat(self.data[key].replace("Z", "+00:00"))
 
-    def get_optional_string(self, key: Union[TikaKey, str]) -> Optional[str]:
+    def get_optional_string(self, key: Union[TikaKey, DublinCoreKey, XmpKey, str]) -> Optional[str]:
         if key not in self.data:
             return None
         return self.data[key]
 
     def __repr__(self) -> str:  # pragma: no cover
         return f"{self.type} response"
-
-
-class ParsedDocument(TikaResponse):
-    """
-    Properties which Tika seems to return for document like
-    parsing
-    """
-
-    @property
-    def content(self) -> Optional[str]:
-        return self.get_optional_string(TikaKey.Content)
-
-    @property
-    def content_length(self) -> Optional[int]:
-        return self.get_optional_int(TikaKey.ContentLength)
-
-    @property
-    def created(self) -> Optional[datetime]:
-        return self.get_optional_datetime(TikaKey.Created)
-
-    @property
-    def modified(self) -> Optional[datetime]:
-        return self.get_optional_datetime(TikaKey.Modified)
-
-    @property
-    def page_count(self) -> Optional[int]:
-        return self.get_optional_int("xmpTPg:NPages")
-
-    @property
-    def character_count(self) -> Optional[int]:
-        return self.get_optional_int("meta:character-count")
-
-    @property
-    def language(self) -> Optional[str]:
-        return self.get_optional_string("language")
-
-    @property
-    def last_author(self) -> Optional[str]:
-        return self.get_optional_string("meta:last-author")
-
-    @property
-    def revision(self) -> Optional[int]:
-        return self.get_optional_int("cp:revision")
-
-
-class ParsedImage(TikaResponse):
-    """
-    Properties which seem to be returned for image like
-    parsing
-    """
-
-
-class ParsedEmail(TikaResponse):
-    """
-    Properties for .eml files
-    """
-
-    @property
-    def subject(self):
-        return self.get_optional_string(TikaKey.Subject)
-
-    @property
-    def title(self):
-        return self.get_optional_string(TikaKey.Title)
-
-    @property
-    def content(self) -> Optional[str]:
-        return self.get_optional_string(TikaKey.Content)
-
-
-KNOWN_CONTENT_TYPES: Final[
-    Dict[str, Union[Type[ParsedDocument], Type[ParsedImage], Type[ParsedEmail], Type[TikaResponse]]]
-] = {
-    "application/vnd.oasis.opendocument.text": ParsedDocument,
-    "application/vnd.openxmlformats-officedocument.wordprocessingml.document": ParsedDocument,
-    "application/vnd.oasis.opendocument.spreadsheet": ParsedDocument,
-    "application/vnd.openxmlformats-officedocument.spreadsheetml.sheet": ParsedDocument,
-    "application/msword": ParsedDocument,
-    "image/png": ParsedImage,
-    "image/jpeg": ParsedImage,
-    "image/webp": ParsedImage,
-    "message/rfc822": ParsedEmail,
-}