+++ CHANGELOG.md
@@ -5,6 +5,18 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.2.0] - 2023-06-26
+
+### Fixed
+
+- Handling of filenames in the `Content-Disposition` header with non-ASCII characters
+
+### Changed
+
+- All endpoints now return a `TikaResponse`, which will have many of the common keys parsed into Python
+  native data types where possible, based on the list [from the Tika wiki](https://cwiki.apache.org/confluence/display/TIKA/Metadata+Overview).
+  If a key is not in the response, the value will be `None`
+
 ## [0.1.1] - 2023-06-18
 
 ### Fixed